requirements.txt
absl-py==2.1.0
aiosqlite==0.20.0
altgraph==0.17.4
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
bs4==0.0.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
comtypes==1.3.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
distro==1.9.0
dm-tree==0.1.8
et-xmlfile==1.1.0
executing==2.0.1
experta==1.9.4
fake-useragent==1.5.1
Flask==3.0.2
fFlask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.47.0
frozendict==1.2
future==1.0.0
gast==0.5.4
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-api-python-client==2.142.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
groq==0.9.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.4
ipykernel==6.28.0
ipynb-py-convert==0.4.6
ipython==8.19.0
iso8601==2.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.5.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.0
jupyter_core==5.5.1
keras==3.0.5
kiwisolver==1.4.5
langchain-core==0.2.37
langchain-groq==0.1.9
langgraph==0.2.15
langgraph-checkpoint==1.0.8
langgraph-checkpoint-sqlite==1.0.1
langsmith==0.1.108
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
mutagen==1.47.0
mysql-connector-python==8.2.0
namex==0.0.7
nest-asyncio==1.5.8
numpy==1.26.3
openai==1.42.0
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.7
packaging==23.2
pandas==2.2.0
parso==0.8.3
pbr==6.0.0
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.1.0
playsound==1.3.0
ply==3.11
pocketsphinx==5.0.3
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.21.12
psutil==5.9.7
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pycryptodome==3.11.0
pycryptodomex==3.19.0
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.5.2
Pygments==2.17.2
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pylatexenc==2.10
pyparsing==3.1.1
pypdf==4.2.0
PyPDF2==3.0.1
pyperclip==1.8.2
pypiwin32==223
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
qiskit==0.45.1
qiskit-terra==0.45.1
requests==2.32.3
rich==13.7.0
rsa==4.9
rustworkx==0.13.2
schema==0.6.7
scikit-learn==1.4.1.post1
scipy==1.11.4
serial==0.0.97
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.6
SpeechRecognition==3.10.1
SQLAlchemy==2.0.28
sr==0.1.0
stack-data==0.6.3
stevedore==5.1.0
sympy==1.12
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.0rc0
tensorflow-intel==2.16.0rc0
termcolor==2.4.0
tf==1.0.0
threadpoolctl==3.3.0
tk==0.1.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.0
typer==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.12
websockets==13.1
Werkzeug==3.0.1
wheel==0.42.0
winregistry==1.1.1
wrapt==1.16.0
yt-dlp==2024.10.22
